@@ -1,107 +1,83 @@
 {
-    "name": "symfony/framework-standard-edition",
-    "license": "MIT",
+    "name": "root/cccsf3",
+    "license": "proprietary",
     "type": "project",
-    "description": "The \"Symfony Standard Edition\" distribution",
     "autoload": {
-        "psr-0": { "": "src/", "SymfonyStandard": "app/" }
+        "psr-4": {
+            "AppBundle\\": "src/AppBundle",
+            "Numa\\": "src/Numa"
+        },
+        "classmap": [
+            "app/AppKernel.php",
+            "app/AppCache.php"
+        ]
     },
-    "repositories": [
-    {
-        "type": "package",
-        "package": {
-            "name": "jquery/jquery",
-            "version": "1.9.1",
-            "dist": {
-                "url": "http://code.jquery.com/jquery-1.9.1.js",
-                "type": "file"
-            }
-        }
-    }
-    ],
-    
+    "autoload-dev": {
+        "psr-4": {
+            "Tests\\": "tests/"
+        },
+        "files": [
+            "vendor/symfony/symfony/src/Symfony/Component/VarDumper/Resources/functions/dump.php"
+        ]
+    },
     "require": {
-        "php": ">=5.3.3",
-        "symfony/symfony": "2.8.0",
-        "doctrine/orm": "~2.2,>=2.2.3",
-        "doctrine/doctrine-bundle": "~1.2",
-        
-         "twig/extensions": "dev-master",
-        "symfony/assetic-bundle": "~2.3",
-        "symfony/swiftmailer-bundle": "~2.3",
-        "symfony/monolog-bundle": "~2.4",
-        "sensio/distribution-bundle": "~3.0.12",
-
-        "incenteev/composer-parameter-handler": "~2.0",
-        "sensio/generator-bundle": "2.3.*",
-        "braincrafted/bootstrap-bundle": "dev-master",
-        "apy/datagrid-bundle": "dev-master",
-        "beberlei/DoctrineExtensions" : "dev-master",
-        "genemu/form-bundle": "2.2.*",
-        "liip/imagine-bundle":"dev-master",
-        "white-october/pagerfanta-bundle": "dev-master",
-        "gregwar/image-bundle": "dev-master",
-        "endroid/qrcode-bundle": "dev-master",
-        "twbs/bootstrap": "dev-master",
-        "leaseweb/memcache-bundle": "*",
-        "leaseweb/api-caller-bundle": "*",
-        "Trsteel/ckeditor-bundle": "~1.6",
-        "jms/serializer-bundle": "^1.1@dev",
-        "friendsofsymfony/rest-bundle": "dev-master",
-        "nelmio/api-doc-bundle": "^2.9@dev",
-        "comur/image-bundle": "1.0.*@dev",
-        "misd/guzzle-bundle": "dev-master",
-        "eightpoints/guzzle-bundle": "^4.4",
-        "leaseweb/guzzle-bundle": "*",
-        "ci/restclientbundle": "^1.0",
-        "guzzle/guzzle": "~3.9",
-        "pugx/autocompleter-bundle": "^1.2",
+        "php": ">=5.5.9",
+        "doctrine/doctrine-bundle": "^1.6",
+        "doctrine/orm": "^2.5",
+        "doctrine/annotations": "1.4.*",
+        "doctrine/dbal": "2.5.4",
+        "incenteev/composer-parameter-handler": "^2.0",
+        "sensio/distribution-bundle": "^5.0.19",
+        "sensio/framework-extra-bundle": "^3.0.2",
+        "symfony/monolog-bundle": "^3.1.0",
+        "symfony/polyfill-apcu": "^1.0",
+        "symfony/swiftmailer-bundle": "^2.3.10",
+        "symfony/symfony": "3.3.*",
+        "twig/twig": "^1.0||^2.0",
+        "symfony/assetic-bundle": "^2.8",
         "knplabs/knp-snappy-bundle": "dev-master",
-        "liip/theme-bundle": "dev-master",
-        "mpdf/mpdf": "dev-master",
+        "white-october/pagerfanta-bundle": "dev-master",
+        "twig/extensions": "dev-master",
         "liuggio/excelbundle": "~2.0",
-        "friendsofsymfony/elastica-bundle": "dev-master",
-        "ruflin/elastica": "3.2.1 as 2.99.0",
-        "sensio/buzz-bundle": "dev-master",
-        "league/oauth1-client": "^1.7",
-        "wheniwork/oauth1-intuit": "^0.1.1",
-        "elliotchance/intuit-quickbooks": "dev-master",
-        "mediafigaro/google-analytics-api-symfony": "^1.0",
-        "google/apiclient": "^2.0",
-        "consolibyte/quickbooks": "dev-master"
+        "mpdf/mpdf": "dev-master",
+        "laurentbrieu/tcpdf": "dev-master",
+        "bitgrave/barcode-bundle": "dev-master",
+        "royopa/fpdf-symfony2": "^1.0",
+        "genemu/form-bundle": "dev-master"
+    },
+    "require-dev": {
+        "sensio/generator-bundle": "^3.0",
+        "symfony/phpunit-bridge": "^3.0"
     },
     "scripts": {
-        "post-install-cmd": [
+        "symfony-scripts": [
             "Incenteev\\ParameterHandler\\ScriptHandler::buildParameters",
             "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::buildBootstrap",
             "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::clearCache",
             "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installAssets",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installRequirementsFile"
+            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installRequirementsFile",
+            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::prepareDeploymentTarget"
+        ],
+        "post-install-cmd": [
+            "@symfony-scripts"
         ],
         "post-update-cmd": [
-            "Incenteev\\ParameterHandler\\ScriptHandler::buildParameters",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::buildBootstrap",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::clearCache",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installAssets",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installRequirementsFile"
+            "@symfony-scripts"
         ]
     },
     "config": {
-        "bin-dir": "bin"
+        "sort-packages": true
     },
-    "minimum-stability": "dev",
     "extra": {
         "symfony-app-dir": "app",
+        "symfony-bin-dir": "bin",
+        "symfony-var-dir": "var",
         "symfony-web-dir": "web",
+        "symfony-tests-dir": "tests",
+        "symfony-assets-install": "relative",
         "incenteev-parameters": {
             "file": "app/config/parameters.yml"
         },
-        "branch-alias": {
-            "dev-master": "2.3-dev"
-        }
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^5.7",
-        "codeception/codeception": "^2.2"
+        "branch-alias": null
     }
 }
